Tratamento de fluxo aprimorado para os métods de atualização, exibição e deleção, para caso as marcas passadas não existam.

+++ app/Http/Controllers/MarcaController.php
@@ -55,7 +55,11 @@ class MarcaController extends Controller
      */
     public function show($id)
     {
-        return $this->marca->find($id);
+        $marca = $this->marca->find($id);
+        if($marca === null){
+            return ['erro' => 'Recurso indisponível!'];
+        }
+        return $marca;
     }
 
     /**
@@ -79,6 +83,9 @@ class MarcaController extends Controller
     public function update(Request $request, $id)
     {
         $marca = $this->marca->find($id);
+        if($marca === null){
+            return ['erro' => 'Impossível realizar a atualização. O recurso solicitado não existe!'];
+        }
         $marca->update($request->all());
         return $marca;
     }
@@ -92,6 +99,9 @@ class MarcaController extends Controller
     public function destroy($id)
     {
         $marca = $this->marca->find($id);
+        if($marca === null){
+            return ['erro' => 'Impossível realizar a exclusão. O recurso solicitado não existe!'];
+        }
         $marca->delete();
         return ['msg' => 'A marca foi removida com sucesso!'];
     }
